Generate ECEG-AH logs excluding key generation
I didn't precompute pk's and sk's and put them in a file,
as it seems I didn't need to. The values didn't take long
to generate and never failed for the entropy reason like
before. I also added a single program to spin up the
client and server in separate processes.