Complete EG-AH and run benchmark tests
The problem here wasn't with the code but with an
assumption about openssl's BN_mod_exp function.
Apparently, this function takes the absolute value
of the argument passed for the exponent and for this
reason equivalent list entries weren't cancelling out.
By taking the exponent with a positive value and then
inverting it we can achieve the same result as raising
to a negative power.