Refactor library to allow for pli protocol specification
This allows us to specify which pli protocol we want to run,
e.g. pli, plica, etc. This also has new logs for all of the existing
implementations. This also runs the EG versions with exponents
that have the bit lengths of the orders of the generators of the
ECEG versions.