Implement rule that no headers include other headers
I like this rule, because I can guarantee that I won't have multiple
includes and it seems like I've avoided circular dependencies for
the time being. That might not always be the case but as of right
now the only includes you see are in the .c files and the headers
have comments detailing the necessary prerequisite includes.